+++ checkpoint/CHANGELOG.md
@@ -195,6 +195,7 @@ Refactored `metadata/loading.py` to allow this change.
 
 ### Added
 
+- Add `assert_tensor_spec_with_default` for input signature testing
 - Add support for loading SafeTensors checkpoints
 - #v1 Add `is_orbax_checkpoint()` method for validation checks
 - #v1 Add custom LeafHandler docs.